+negative indexing

+++ src/rotta_rs/arrayy/operation/math/matmul.rs
@@ -15,7 +15,12 @@ pub fn matmul_2d(arr_a: &Arrayy, arr_b: &Arrayy) -> Arrayy {
             for coll in 0..*o {
                 let mut sum = 0.0;
                 for i in 0..*n {
-                    sum = sum + (arr_a.index(vec![row, i]) * arr_b.index(vec![i, coll])).value[0];
+                    sum =
+                        sum +
+                        (
+                            arr_a.index(vec![row as i32, i as i32]) *
+                            arr_b.index(vec![i as i32, coll as i32])
+                        ).value[0];
                 }
                 vector.push(sum);
             }
@@ -68,7 +73,7 @@ pub fn matmul_nd(arr_a: &Arrayy, arr_b: &Arrayy) -> Arrayy {
                 } else {
                     index[d] += 1;
 
-                    if index[d] >= shape_a[d] {
+                    if index[d] >= (shape_a[d] as i32) {
                         index.pop();
 
                         if d == 0 {
