+++ src/rotta_rs/tensor/function/sum_axis.rs
@@ -1,6 +1,6 @@
 use crate::rotta_rs::{ arrayy::sum_axis_arr, to_shape, Arrayy, BackwardLabel, NodeType, Tensor };
 
-pub fn sum_axis(x: &Tensor, d: usize) -> Tensor {
+pub fn sum_axis(x: &Tensor, d: i32) -> Tensor {
     let array = x.value();
     let tensor = Tensor::from_arrayy(sum_axis_arr(&array, d));
     tensor.update_parent(vec![x.node.clone()]);
@@ -9,12 +9,12 @@ pub fn sum_axis(x: &Tensor, d: usize) -> Tensor {
     tensor
 }
 
-pub fn sum_axis_keep_dim(x: &Tensor, d: usize) -> Tensor {
+pub fn sum_axis_keep_dim(x: &Tensor, d: i32) -> Tensor {
     let array = x.value();
 
     let sum = sum_axis_arr(&array, d);
     let mut keep_dim = sum.shape.clone();
-    keep_dim.insert(d, 1);
+    keep_dim.insert(d as usize, 1);
 
     let tensor = Tensor::from_vector(keep_dim, sum.value);
     tensor.update_parent(vec![x.node.clone()]);
@@ -23,14 +23,14 @@ pub fn sum_axis_keep_dim(x: &Tensor, d: usize) -> Tensor {
     tensor
 }
 
-pub fn d_sum_axis(x: &NodeType, d: usize, keep_dim: bool, grad: &Arrayy) {
+pub fn d_sum_axis(x: &NodeType, d: i32, keep_dim: bool, grad: &Arrayy) {
     let mut x = x.lock().unwrap();
 
     if x.requires_grad {
         if !keep_dim {
             let ones = Arrayy::ones(x.value.shape.clone());
             let mut new_shape = grad.shape.clone();
-            new_shape.insert(d, 1);
+            new_shape.insert(d as usize, 1);
 
             let d = ones * to_shape(grad, new_shape);
             x.add_grad(d);
